refactor(helper): move app.constant into helper.constant
