+++ src/biomodals/app/bioinfo/paddleocr_app.py
@@ -9,8 +9,8 @@ from pathlib import Path
 import modal
 
 from biomodals.app.config import AppConfig
-from biomodals.app.constant import MODEL_VOLUME
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MODEL_VOLUME
 from biomodals.helper.shell import package_outputs, softlink_dir
 
 ##########################################
