+++ src/biomodals/app/design/boltzgen_app.py
@@ -17,8 +17,8 @@ import modal
 import orjson
 
 from biomodals.app.config import AppConfig
-from biomodals.app.constant import MAX_TIMEOUT, MODEL_VOLUME
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MAX_TIMEOUT, MODEL_VOLUME
 from biomodals.helper.shell import (
     package_outputs,
     run_command,
