+++ src/biomodals/app/fold/abcfold2_app.py
@@ -23,8 +23,8 @@ from pathlib import Path
 import modal
 
 from biomodals.app.config import AppConfig
-from biomodals.app.constant import MODEL_VOLUME
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MODEL_VOLUME
 from biomodals.helper.shell import package_outputs
 from biomodals.helper.web import download_files
