+++ src/biomodals/app/score/af3score_app.py
@@ -23,8 +23,8 @@ from tempfile import TemporaryDirectory, mkdtemp
 import modal
 
 from biomodals.app.config import AppConfig
-from biomodals.app.constant import MODEL_VOLUME
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MODEL_VOLUME
 from biomodals.helper.shell import (
     copy_files,
     run_command,
