+++ src/biomodals/workflow/ppiflow_workflow.py
@@ -37,8 +37,8 @@ from modal.volume import FileEntryType
 from uniaf3.schema.alphafold3 import AF3Config, AF3Protein, AF3SequenceEntry
 
 from biomodals.app.config import AppConfig
-from biomodals.app.constant import MAX_TIMEOUT
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MAX_TIMEOUT
 from biomodals.helper.shell import package_outputs, sanitize_filename
 
 ##########################################
