+++ src/biomodals/workflow/shortmd_workflow.py
@@ -19,8 +19,8 @@ from pathlib import Path, PurePosixPath
 import modal
 
 from biomodals.app.bioinfo import gromacs_app
-from biomodals.app.constant import MAX_TIMEOUT
 from biomodals.helper import patch_image_for_helper
+from biomodals.helper.constant import MAX_TIMEOUT
 from biomodals.helper.shell import sanitize_filename
 from biomodals.schema import (
     AppConfig,
